==> 4_Sentiment analysis.ipynb <==
### 本篇代码用于计算每个文件的sentiment score

import pandas as pd
import numpy as np
pd.options.mode.chained_assignment = None
import pysentiment as ps

df = pd.read_pickle("Pickles\\processed_data.pkl")
df.head()

lm = ps.LM()#use the Loughran and McDonald dictionary

#写一个函数利用字典提取情感值
def sentiment_score(text):
    dic = lm.get_score(text)
    return dic['Polarity']

用Loughran and McDonald dictionary提取情感

df['sentiment'] = df['processed_text'].apply(sentiment_score)

df.head()

df['sentiment'].describe()

df.to_pickle("Pickles\\processed_data.pkl")
