==> 5_Model construction_benchmark model.ipynb <==
### 本篇代码用于建立对照模型——SVM


from sklearn.metrics import roc_auc_score
from keras.utils import np_utils,plot_model, multi_gpu_model
from IPython.display import SVG
from keras.utils.vis_utils import model_to_dot

from sklearn.model_selection import StratifiedShuffleSplit #sklearn是python自带的机器学习库
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import MultiLabelBinarizer
from sklearn.preprocessing import StandardScaler

from sklearn.feature_extraction.text import CountVectorizer
from sklearn import svm
from sklearn.svm import SVC, LinearSVC
from sklearn.metrics import classification_report
from sklearn.metrics import f1_score
from sklearn.metrics import accuracy_score 
from sklearn.metrics import confusion_matrix

import pandas as pd
import scipy
from scipy import sparse

df = pd.read_pickle("Pickles/processed_data.pkl")
#df = pd.read_csv("Data/lemmatized_text.csv")

df.head()

#### 把项目变量变成n-1个哑变量

mlb = MultiLabelBinarizer()
df = df.join(pd.DataFrame(mlb.fit_transform(df.pop('items')),columns=mlb.classes_,),sort=False,how="left")

df.head()

#### 定义一个函数将文本连接起来方便后面向量化，为SVM建模做准备 （预处理时文本被tokenize）


def combine_text(text):
    combined_text = []
    combined_text = ' '.join(str(x) for x in text)
    return combined_text

df['processed_con_text'] = df['processed_text'].apply(combine_text)

#### 把其他类别变量变成哑变量，把label和数据分开

# 把label和数据分开，所有类别变量都变成哑变量
cols = ['GICS Sector','vix','rm_week','rm_month', 'rm_qtr', 'rm_year','sentiment']
cols.extend(list(mlb.classes_))
X = df[cols]
docs = df['processed_text']

y = df['signal']

# Get Dummies

#docs = tokenize_and_pad(docs)
X = pd.get_dummies(columns = ['GICS Sector'],prefix="sector",data=X)
#y = pd.get_dummies(columns=['signal'],data=y)

#aux_shape = len(X.columns)

检查是否正确

len(y)

len(X)

len(docs)

X.head()

#### 将连续变量标准化


cont_features = ['vix','rm_week','rm_month', 'rm_qtr', 'rm_year']

x_scaler = StandardScaler()
X[cont_features] = x_scaler.fit_transform(X[cont_features])



X.head()

#### 将数据划分为training,testing


X_train, X_test, y_train, y_test, docs_train, docs_test = train_test_split(X, y,docs,
                                                    stratify=y, 
                                                    test_size=0.2,
                                                    random_state = 20)
#X_validation, X_test, y_validation, y_test, docs_validation, docs_test = train_test_split(X_test_va, y_test_va,docs_test_va,
 #                                                   stratify=y_test_va, 
  #                                                  test_size=0.5,
   #                                                 random_state = 20)

y.head()

#### oversample数据调整训练集中的样本不均衡问题

def oversample(X,docs,y):
    # Get number of rows with imbalanced class
    target = y.sum().idxmax()#得到y中数量最多的对象，up
    n = y[target].sum()#计数
    # identify imbalanced targets
    imbalanced = y.drop(target,axis=1) #扔掉up，保留另外两列
    #For each target, create a dataframe of randomly sampled rows, append to list
    #随机从另外两列中抽取一些数据进行复制
    append_list =  [y.loc[y[col]==1].sample(n=n-y[col].sum(),replace=True,random_state=20) for col in imbalanced.columns]
    append_list.append(y)
    #复制时他们的index也复制了，所以可以用来匹配X
    y = pd.concat(append_list,axis=0)
    # match y indexes on other inputs
    X = X.loc[y.index] #注意这里是根据两个表的索引来匹配的，一旦索引除了问题就会出错
    docs = docs.loc[y.index]
    assert (y.index.all() == X.index.all() == docs.index.all())
    return X,docs,y

#调整样本代表性
X_train_svm, docs_train_svm, y_train_svm = oversample(X_train, docs_train, y_train)

# Save data

X_train_svm.to_pickle("Pickles/X_train_svm.pkl")
X_test_svm.to_pickle("Pickles/X_test_svm.pkl")
#X_validation.to_pickle("Pickles/X_validation.pkl")

y_train_svm.to_pickle("Pickles/y_train_svm.pkl")
y_test_svm.to_pickle("Pickles/y_test_svm.pkl") 

docs_train_svm.to_pickle("Pickles/docs_train_svm")
docs_test_svm.to_pickle("Pickles/docs_test_svm")


#### 导入模型进行训练

先定义一个函数来计算多分类问题的auc

def auc_svm(test,pred,average='macro'):
    lb=LabelBinarizer()
    lb.fit(test)
    test=lb.transform(test)
    pred=lb.transform(pred)
    return roc_auc_score(test,pred,average=average)

先用文本数据进行训练和测试

basicvectorizer = CountVectorizer(ngram_range=(1,1))
svm_vec_train = basicvectorizer.fit_transform(docs_train_svm)
print(svm_vec_train.shape)

svm_vec_test = basicvectorizer.transform(docs_test_svm)
print(svm_vec_test.shape)

svm_model = svm.LinearSVC(C=1,max_iter=10000)
#svm_model_trained = svm_model.fit([svm_vec_train,X_train_svm], y_train_svm)
svm_model_trained = svm_model.fit(svm_vec_train, y_train_svm)

其实从这里就可以看出这个文本代表方法的缺陷——模型的迭代设置到了10000（默认值的10倍），仍然无法收敛

predictions = svm_model_trained.predict(svm_vec_test)

print (classification_report(y_test_svm, predictions))


采用macro的方法，因为auc的函数也是采用的这个指标。上面这个报告里的precision就是每一种分类里面的accuracy，他们的平均就是整体的accuracy.

auc_svm(y_test_svm, predictions)

用非文本数据进行训练和测试

svm_model = svm.LinearSVC(C=1,max_iter=10000)
#svm_model_trained = svm_model.fit([svm_vec_train,X_train_svm], y_train_svm)
svm_model_trained = svm_model.fit(X_train_svm, y_train_svm)

predictions = svm_model_trained.predict(X_test_svm)

print (classification_report(y_test_svm, predictions))

auc_svm(y_test_svm, predictions)

用整体数据进行训练和测试

train_all = sparse.hstack((svm_vec_train, X_train_svm)).tocsr()#将文本向量和普通变量合在一起

test_all = sparse.hstack((svm_vec_test, X_test_svm)).tocsr()

svm_model = svm.LinearSVC(C=1,max_iter=10000)
#svm_model_trained = svm_model.fit([svm_vec_train,X_train_svm], y_train_svm)
svm_model_trained = svm_model.fit(train_all, y_train_svm)

predictions = svm_model_trained.predict(test_all)

print (classification_report(y_test_svm, predictions))


auc_svm(y_test_svm,predictions)
